requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.19
apscheduler==3.10.4
pandas==2.0.3
beautifulsoup4==4.12.2
pytz==2023.3